requirements.txt
boto==2.49.0
email-validator
flask==2.2.5
flask-login==0.6.2
flask-migrate==4.0.4
flask-minify==0.42
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
gitpython
jinja2==3.1.2
nats-py==2.1.7
python-dotenv==1.0.0
wtforms==3.0.1
Werkzeug==2.2.2
gunicorn==20.1.0

